replace script generating colony building scripts with a script that just creates colony buildings when run by parser
